+++ sharding-core/src/main/java/io/shardingsphere/core/executor/sql/execute/result/DistinctQueryResult.java
@@ -22,7 +22,7 @@ import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Multimap;
-import io.shardingsphere.core.executor.sql.execute.result.row.QueryRow;
+import io.shardingsphere.core.executor.sql.execute.row.QueryRow;
 import io.shardingsphere.core.merger.QueryResult;
 import lombok.AccessLevel;
 import lombok.Getter;
